feat : AlarmData Domain계층 생성 및 Di 적용

+++ data/src/main/java/com/grusie/data/database/AlarmDao.kt
@@ -16,7 +16,7 @@ interface AlarmDao {
     suspend fun delete(alarm: AlarmData)
 
     @Query("SELECT * FROM alarm_table WHERE id = :id")
-    suspend fun getAlarmById(id: Int): AlarmData
+    suspend fun getAlarmById(id: Long): AlarmData
 
     @Query("SELECT * FROM alarm_table ORDER BY time ASC")
     fun getAllAlarms(): List<AlarmData>
